% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionChina.r
\name{mapChina_fill_province}
\alias{mapChina_fill_province}
\title{mapChinaFillProvince.}
\usage{
mapChina_fill_province(
  data,
  label,
  colorTrans = NULL,
  title = "",
  highColor = "red",
  lowColor = "yellow",
  naColor = "white",
  axis.text.size = 9,
  legend.text.size = 10,
  legend.title.size = 11
)
}
\description{
draw China map in province.
}
