% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionChina.r
\name{mapChina_pie_province}
\alias{mapChina_pie_province}
\title{mapChinaPieProvince.}
\usage{
mapChina_pie_province(
  data,
  title = "",
  pieName = NULL,
  axis.text.size = 9,
  dataCol = NULL,
  piePalette = "Set2",
  pieSize = 1,
  NameCol = 1
)
}
\description{
draw China map in province.
}
