% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionChina.r
\name{mapChina_point}
\alias{mapChina_point}
\title{mapChinaPoint.}
\usage{
mapChina_point(
  data,
  title = "",
  colorby = NULL,
  shapeby = NULL,
  colorTrans = "log10",
  size = 2,
  shape = 19,
  color = "black",
  axis.text.size = 9
)
}
\description{
draw China map in Points.
}
