% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionWorld.r
\name{mapWorld_fill}
\alias{mapWorld_fill}
\title{mapWorld_fill.}
\usage{
mapWorld_fill(
  data,
  label,
  colorTrans = NULL,
  style = "round",
  title = "",
  scale = "medium",
  highColor = "red",
  lowColor = "yellow",
  naColor = "white",
  axis.text.size = 9,
  legend.text.size = 10,
  legend.title.size = 11
)
}
\description{
draw World map in country.
}
