% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionWorld.r
\name{mapWorld_point}
\alias{mapWorld_point}
\title{mapWorld_point.}
\usage{
mapWorld_point(
  data,
  title = "",
  inputType = "coordinate",
  bwmap = F,
  style = "round",
  alpha = 0.8,
  colorby = NULL,
  shapeby = NULL,
  jitter.factor = 0.002,
  scale = "medium",
  colorTrans = "log10",
  size = 2,
  shape = 20,
  color = "black",
  axis.text.size = 9,
  jitter = F,
  point.mapcolors = "Set1"
)
}
\description{
draw Points in World map.
}
